GH-435: data retention sanitization
dds-session: Added: A data retention sanitization. Not running sessions older than the specified number of days ("server.data_retention") are auto deleted. (GH-435)
dds-user-defaults: Modified: Bumo the version to 0.5
dds-user-defaults: Added: A "server.data_retention" configuration key. (GH-435)

+++ dds-protocol-lib/src/ConnectionManagerImpl.h
@@ -88,7 +88,7 @@ namespace dds
                     A* pThis = static_cast<A*>(this);
                     pThis->_start();
 
-                    const float maxIdleTime =
+                    const double maxIdleTime =
                         user_defaults_api::CUserDefaults::instance().getOptions().m_server.m_idleTime;
 
                     CMonitoringThread::instance().start(maxIdleTime,
